transfer 2 contest helper functions
