+++ libs/evm-protocols/src/common-protocol/contractHelpers/contest.ts
@@ -1,5 +1,9 @@
 import { BigNumber } from '@ethersproject/bignumber';
+import { feeManagerAbi } from '@hicommonwealth/evm-protocols';
 import { ZERO_ADDRESS } from '@hicommonwealth/shared';
+import Web3 from 'web3';
+import { recurringContestAbi } from '../../abis/recurringContestAbi';
+import { singleContestAbi } from '../../abis/singleContestAbi';
 
 export const getTotalContestBalance = async (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -59,4 +63,135 @@ export const getTotalContestBalance = async (
       : BigNumber.from(balanceResults[0]);
 
   return BigNumber.from(balance).toString();
+};
+
+/**
+ * Gets relevant contest state information
+ * @param rpcNodeUrl the rpc node url
+ * @param contest the address of the contest
+ * @param oneOff boolean indicating whether this is a recurring contest - defaults to false (recurring)
+ * @returns Contest Status object
+ */
+export const getContestStatus = async (
+  rpcNodeUrl: string,
+  contest: string,
+  oneOff: boolean = false,
+): Promise<{
+  startTime: number;
+  endTime: number;
+  contestInterval: number;
+  lastContentId: string;
+}> => {
+  const web3 = new Web3(rpcNodeUrl);
+  const contestInstance = new web3.eth.Contract(
+    oneOff ? singleContestAbi : recurringContestAbi,
+    contest,
+  );
+
+  const promise = await Promise.all([
+    contestInstance.methods.startTime().call(),
+    contestInstance.methods.endTime().call(),
+    oneOff
+      ? contestInstance.methods.contestLength().call()
+      : contestInstance.methods.contestInterval().call(),
+    contestInstance.methods.currentContentId().call(),
+  ]);
+
+  return {
+    startTime: Number(promise[0]),
+    endTime: Number(promise[1]),
+    contestInterval: Number(promise[2]),
+    lastContentId: String(promise[3]),
+  };
+};
+
+/**
+ * Get the total balance of a given contest
+ * @param rpcNodeUrl the rpc node url
+ * @param contest the address of contest to get the balance of
+ * @param oneOff boolean indicating whether this is a recurring contest - defaults to false (recurring)
+ * @returns a numeric contest balance of the contestToken in wei(ie / 1e18 for decimal value)
+ */
+export const getContestBalance = async (
+  rpcNodeUrl: string,
+  contest: string,
+  oneOff: boolean = false,
+): Promise<string> => {
+  const web3 = new Web3(rpcNodeUrl);
+
+  const contestInstance = new web3.eth.Contract(
+    oneOff ? singleContestAbi : recurringContestAbi,
+    contest,
+  );
+
+  return await getTotalContestBalance(
+    contestInstance,
+    contest,
+    web3,
+    feeManagerAbi,
+    oneOff,
+  );
+};
+
+/**
+ * Gets vote and more information about winners of a given contest
+ * @param rpcNodeUrl the rpc node url
+ * @param contest the address of the contest
+ * @param contestId the id of the contest for data within the contest contract.
+ * No contest id will return current winners
+ * @param oneOff boolean indicating whether this is a recurring contest - defaults to false (recurring)
+ * @returns ContestScores object containing eqaul indexed content ids, addresses, and votes
+ */
+export const getContestScore = async (
+  rpcNodeUrl: string,
+  contest: string,
+  contestId?: number,
+  oneOff: boolean = false,
+): Promise<{
+  scores: {
+    winningContent: string;
+    winningAddress: string;
+    voteCount: string;
+  }[];
+  contestBalance: string;
+}> => {
+  const web3 = new Web3(rpcNodeUrl);
+  const contestInstance = new web3.eth.Contract(
+    oneOff ? singleContestAbi : recurringContestAbi,
+    contest,
+  );
+
+  const contestData = await Promise.all([
+    contestId
+      ? contestInstance.methods.getPastWinners(contestId).call()
+      : contestInstance.methods.getWinnerIds().call(),
+    getContestBalance(rpcNodeUrl, contest, oneOff),
+  ]);
+
+  const winnerIds: string[] = contestData[0] as string[];
+
+  if (winnerIds.length == 0) {
+    throw new Error(
+      `getContestScore ERROR: No winners found for contest ID (${contestId}) on contest address: ${contest}`,
+    );
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const votePromises: any[] = [];
+  winnerIds.forEach((w) => {
+    votePromises.push(contestInstance.methods.content(w).call());
+  });
+
+  const contentMeta = await Promise.all(votePromises);
+
+  return {
+    scores: winnerIds.map((v, i) => {
+      return {
+        winningContent: v,
+        winningAddress: contentMeta[i]['creator'],
+        voteCount: contentMeta[i]['cumulativeVotes'],
+      };
+    }),
+    contestBalance: contestData[1],
+  };
 };
