+++ libs/model/src/contest/Contests.projection.ts
@@ -4,6 +4,8 @@ import {
   ChildContractNames,
   EvmEventSignatures,
   commonProtocol as cp,
+  getContestScore,
+  getContestStatus,
   getTokenAttributes,
 } from '@hicommonwealth/evm-protocols';
 import { config } from '@hicommonwealth/model';
@@ -14,7 +16,6 @@ import { z } from 'zod';
 import { models } from '../database';
 import { mustExist } from '../middleware/guards';
 import { EvmEventSourceAttributes } from '../models';
-import * as protocol from '../services/commonProtocol';
 import { getWeightedNumTokens } from '../services/stakeHelper';
 import {
   decodeThreadContentUrl,
@@ -82,7 +83,7 @@ async function updateOrCreateWithAlert(
     true,
   );
 
-  const { startTime, endTime } = await protocol.contestHelper.getContestStatus(
+  const { startTime, endTime } = await getContestStatus(
     url,
     contest_address,
     isOneOff,
@@ -223,13 +224,12 @@ export async function updateScore(contest_address: string, contest_id: number) {
         `Chain node url not found on contest ${contest_address}`,
       );
 
-    const { scores, contestBalance } =
-      await protocol.contestHelper.getContestScore(
-        details.url,
-        contest_address,
-        undefined,
-        oneOff,
-      );
+    const { scores, contestBalance } = await getContestScore(
+      details.url,
+      contest_address,
+      undefined,
+      oneOff,
+    );
 
     const prizePool = BigNumber.from(contestBalance)
       .mul(oneOff ? 100 : details.prize_percentage)
